+++ build.gradle
@@ -35,7 +35,6 @@ apply plugin: 'signing'
 apply plugin: 'maven-publish'
 
 dependencies {
-
   testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.2'
   testImplementation group: 'com.ginsberg', name: 'junit5-system-exit', version: '1.1.2'
   testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.2'
