+++ src/main/java/org/rossonet/command/parser/action/ErrorAction.java
@@ -1,6 +1,8 @@
 package org.rossonet.command.parser.action;
 
+import org.rossonet.command.parser.exception.CommandParserException;
 import org.rossonet.command.parser.option.Option;
+import org.rossonet.command.parser.option.token.TokenMatch;
 
 public interface ErrorAction {
