+++ src/main/java/org/rossonet/command/parser/option/token/TokenBuilder.java
@@ -1,4 +1,4 @@
-package org.rossonet.command.parser.option;
+package org.rossonet.command.parser.option.token;
 
 public interface TokenBuilder<TOKEN extends Token> {
